remove duplicate code part 2
